move stories to components
